+++ app-classic/src/components/Header.svelte
@@ -12,25 +12,26 @@
   const identicon = $derived(avatarSVG(userAlias, 52))
 </script>
 
-<header class="min-h-[var(--header-h)] ps-4 pt-1 pb-1 flex items-start justify-between --border">
+<header class="min-h-[var(--header-h)] px-4 pt-1 pb-1 flex items-center justify-between --border-b --bg-white">
   <div class="flex-none lg:hidden">
     <label for="sidebar-drawer" aria-label="open sidebar" class="btn btn-square btn-ghost btn-sm">
       <Menu class="size-5" />
     </label>
   </div>
 
-  <div class="flex-none py-1">
-    <div class="text-2xl text-indigo-900 font-mono font-bold">
-      <img src="./myown-logo.svg" alt="My Own Logo" class="w-auto h-[48px]" />
+  <div class="flex-none">
+    <div class="text-lg text-gray-900 font-mono font-bold">
+      <span>Biz/Family</span>
+      <!-- <img src="./myown-logo.svg" alt="My Own Logo" class="w-auto h-[48px]" /> -->
     </div>
   </div>
 
   <div class="flex items-center justify-end">
-    <div class="dropdown dropdown-end me-2">
+    <div class="dropdown dropdown-end">
       <div
         tabindex="0"
         role="button"
-        class="btn btn-ghost btn-circle size-11 p-0 overflow-hidden border-0"
+        class="btn btn-ghost btn-circle size-9 overflow-hidden border-0"
         title={userAlias}
         aria-label="Open profile menu"
         >
@@ -57,7 +58,7 @@
       </ul>
     </div>
 
-    <a href={"#"}
+    <!-- <a href={"#"}
       class={
         "bg-white h-9 w-10 flex items-center justify-center border-l border-gray-300 rounded-l-md ps-1"+
         "  border border-gray-300 hover:bg-gray-300"
@@ -66,6 +67,6 @@
       title="Ask me"
       >
       <Sparkles class="size-5 text-orange-600" />
-    </a>
+    </a> -->
   </div>
 </header>